test(fuzz): add structured and raw fuzzing for binary import
- Add `arbitrary` dependency and `fuzzing` feature to `Cargo.toml`.
- Implement `Arbitrary` for `HVec10240`, `Concept`, and all binary export structures.
- Add `binary_payload_logic` fuzz target for structured fuzzing of the internal logic.
- Simplify `binary_import` fuzz target to focus on raw byte/deserialization fuzzing.
- Refactor framework to support `import_binary_from_bytes` for efficient fuzzing.
