+++ src/cli/commands/history.rs
@@ -1,7 +1,7 @@
 //! History command for listing concept version history.
 
-use super::{create_framework_with_namespace, run_get, run_rollback, validate_concept_id};
-use crate::cli::args::{GetArgs, HistoryArgs, OutputFormat, RollbackArgs};
+use super::{create_framework_with_namespace, validate_concept_id};
+use crate::cli::args::{HistoryArgs, OutputFormat};
 use crate::cli::error::{CliError, Result};
 use colored::Colorize;
 use std::path::Path;
@@ -15,35 +15,6 @@ pub async fn run_history(
 ) -> Result<()> {
     validate_concept_id(&args.concept_id)?;
 
-    // Dispatch to get version if --version is specified
-    if let Some(version) = args.version {
-        return run_get(
-            GetArgs {
-                namespace: args.namespace,
-                concept_id: args.concept_id,
-                version: Some(version),
-            },
-            db_path,
-            format,
-        )
-        .await;
-    }
-
-    // Dispatch to rollback if --rollback is specified
-    if let Some(to_version) = args.rollback {
-        return run_rollback(
-            RollbackArgs {
-                namespace: args.namespace,
-                concept_id: args.concept_id,
-                to: to_version,
-                confirm: args.confirm,
-            },
-            db_path,
-            format,
-        )
-        .await;
-    }
-
     let framework = create_framework_with_namespace(db_path, &args.namespace).await?;
     let versions = framework
         .list_versions(&args.concept_id)
@@ -54,11 +25,8 @@ pub async fn run_history(
         OutputFormat::Json => {
             println!(
                 "{}",
-                serde_json::to_string(&serde_json::json!({
-                    "concept_id": args.concept_id,
-                    "versions": versions
-                }))
-                .map_err(|e| CliError::Output(format!("failed to serialize versions: {e}")))?
+                serde_json::to_string(&versions)
+                    .map_err(|e| CliError::Output(format!("failed to serialize versions: {e}")))?
             );
         }
         OutputFormat::Table => {
